+++ src/main/java/net/raphimc/noteblocklib/format/txt/model/TxtNote.java
@@ -21,23 +21,23 @@ import java.util.Objects;
 
 public class TxtNote {
 
-    private byte instrument;
-    private byte key;
+    private int instrument;
+    private int key;
 
-    public byte getInstrument() {
+    public int getInstrument() {
         return this.instrument;
     }
 
-    public TxtNote setInstrument(final byte instrument) {
+    public TxtNote setInstrument(final int instrument) {
         this.instrument = instrument;
         return this;
     }
 
-    public byte getKey() {
+    public int getKey() {
         return this.key;
     }
 
-    public TxtNote setKey(final byte key) {
+    public TxtNote setKey(final int key) {
         this.key = key;
         return this;
     }
